# api_openapi.yml
openapi: 3.0.3
info:
  title: SalmanAid API
  description: API for Rumah Amal Salman Student Loan Management System.
  version: 1.0.0
servers:
  - url: localhost:3001/api
    description: Base URL for the API
paths:
...
  /attachments/{attachmentId}:
    delete:
      summary: Delete document
      security:
        - nextAuth: []
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          description: Forbidden
        '404':
          description: Attachment not found
...
  /donors/distributions:
    get:
      summary: Get donor distributions tracking breakdown
      security:
        - nextAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: donorId
          in: query
          description: Filter by donor UUID (Admins can see any, Donors only their own)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DistributionRecord'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
components:
  securitySchemes:
    nextAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    DistributionRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loanId:
          type: string
          format: uuid
        beneficiaryName:
          type: string
        description:
          type: string
        allocatedAmount:
          type: number
        allocatedAt:
          type: string
          format: date-time